requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo
pymongo[srv]==3.12
scikit-learn
#dill
mlflow
PyYAML 
#-e .